requirements.txt
ckanapi>=4.6
python-dateutil>=2.8
requests>=2.25